Open-ECG-Digitizer/requirements.txt
ipython
matplotlib
numpy
pillow
scikit-image
scipy
torch
torchaudio
torchvision
tqdm
types-tqdm
types-setuptools
yacs
ray[data,train,tune]
tensorboard
networkx
types-networkx
scikit-learn
transformers
plotly
ipywidgets
opencv-python
torch-tps
